refactor(index.js):  removed install script from index
